% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowSorted.CordBloodCombined.450K.R
\docType{data}
\name{FlowSorted.CordBloodCombined.450K}
\alias{FlowSorted.CordBloodCombined.450K}
\title{FlowSorted.CordBloodCombined.450K}
\format{A class: RGChannelSet, dimensions: 575130 289}
\usage{
FlowSorted.CordBloodCombined.450K
}
\description{
The FlowSorted.CordBloodCombined.450K package contains data derived from 
Illumina HumanMethylation450K and Illumina HumanMethylationEPIC DNA 
methylation microarrays (Gervin K, Salas LA et al. 2018), consisting of 263 
blood cell references and 26 umbilical cord blood samples, formatted as an 
RGChannelSet object for integration and normalization using most of the 
existing Bioconductor packages.

This package contains cleaned data from four different umbilical cord blood 
references similar to the FlowSorted.CordBlood.450K package consisting of 
data from umbilical cord blood samples generated from healthy newoborns. 
However, when using the cleaned dataset (eliminating potential cell cross 
contamination) and using the IDOL procedure compared to minfi estimates the 
cell type composition obtained through FlowSorted.CordBlood.450k package were
less precise and biased compared to actual cell counts. Hence, this package 
consists of appropriate data for deconvolution of umbilical cord blood 
samples used in for example EWAS relying in both 450K and EPIC technology.

Researchers may find this package useful as these samples represent different 
cellular populations ( T lymphocytes (CD4+ and CD8+), B cells (CD19+), 
monocytes (CD14+), NK cells (CD56+) and Granulocytes of cell sorted umbilical
 cord blood. The estimates were contrasted versus FACS proportions in 22 
 samples, and validated using 197 umbilical cord blood samples. These data 
 can be integrated with the minfi Bioconductor package to estimate cellular 
 composition in users' umbilical cord blood Illumina 450K and EPIC samples 
 using a modified version of the algorithm constrained projection/quadratic 
 programming described in Houseman et al. 2012. However, for more accurate 
 estimations we suggests that the user prefers IDOL over minfi automatic 
 estimations, using the function estimateCellCounts2 from the package 
 FlowSorted.Blood.EPIC which allows using customized sets of probes from 
 IDOL.
}
\examples{
# Explore the reference library
library(ExperimentHub)
hub <- ExperimentHub()
query(hub, "FlowSorted.CordBloodCombined.450K")
FlowSorted.CordBloodCombined.450K <- hub[["EH1136"]]#Add the correct hub
FlowSorted.CordBloodCombined.450K

}
\seealso{
References \enumerate{
\item LA Salas et al. (2018). \emph{An optimized library for 
reference-based deconvolution of whole-blood biospecimens assayed using the 
Illumina HumanMethylationEPIC BeadArray}. Genome Biology 19, 64. doi:
10.1186/s13059-018-1448-7.
\item DC Koestler et al. (2016). \emph{Improving cell mixture deconvolution 
by identifying optimal DNA methylation libraries (IDOL)}. BMC bioinformatics.
17, 120. doi: 10.1186/s12859-016-0943-7.
\item EA Houseman et al. (2012) \emph{DNA methylation arrays as surrogate
measures of cell mixture distribution}. BMC Bioinformatics 13, 86.
doi:10.1186/1471-2105-13-86.
\item \pkg{minfi} package for tools for estimating cell type
composition in blood using these data
}
}
\keyword{datasets}
